requirements.txt
streamlit>=1.28
openai>=1.10.0
google-genai>=1.73.0
pillow>=10.0
requests>=2.31
numpy<2
moviepy==1.0.3
imageio-ffmpeg==0.6.0
pydantic>=1.10
supabase>=2.0.0
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
mcp>=1.0.0
fal-client>=0.4.0
